Fix registration for Pleroma instances
When making the API call to register, Mastodon would return the
`redirect_uri` field while Pleroma wouldn't. Since the API documentation suggest using `urn:ietf:wg:oauth:2.0:oob` as
a default of sorts, we can use this to handle the difference between instances.

`redirect_uri` is now an Option field, so Serde will not be upset if it is missing, and then if missing the default will be used.

Related: #32